+++ lib/scss_lint/linter/color_variable.rb
@@ -23,6 +23,13 @@ module SCSSLint
         .each   { |_, color| record_lint(node, color) }
     end
 
+    def visit_comment(_node)
+      # Don't lint children. Sass multiline comments (/*...*/) are actually
+      # rendered in code and thus allow variable interpolation. Unfortunately,
+      # the Sass parser returns bad source ranges for interpolation in these
+      # comments, so it's easiest to just ignore them.
+    end
+
   private
 
     def record_lint(node, color)